Attempt to migrate Atomic and test to "Allen-style"
Default CopCore build to target CUDA for now, as CPU targetting requires
more work in AdePT.

Use CopCore header and macros to annotate AdePT Atomic member functions
for host/device and target device. Update test_atomic program to unify
compilation.

NB: This is known to fail compilation even in CUDA target with errors:

.../AdePT.git/base/inc/AdePT/Atomic.h(33):
error: calling a __device__ function("load") from a __host__ __device__
function("Atomic_t") is not allowed

.../AdePT.git/base/inc/AdePT/Atomic.h(37):
error: calling a __device__ function("load") from a __host__ __device__
function("operator=") is not allowed

The source of these errors is not clear as yet. The VecCore macros
*should* expand to the same as the CopCore/Allen ones when targetting
CUDA so why compilation works with VecCore and not CopCore.

It was noted that the device-only functions in Atomic should be
host-device, but they should also be device-only when built with
VecCore.

Commit amended to also take Andrei's latest updates to CopCore
atomic definitions during rebase.